Conflicts:
	app/org/sagebionetworks/bridge/services/SessionUpdateService.java

+++ app/org/sagebionetworks/bridge/services/SurveyService.java
@@ -12,12 +12,13 @@ import java.util.Set;
 import java.util.function.BiPredicate;
 import java.util.stream.Collectors;
 
-import org.apache.commons.lang3.StringUtils;
 import org.sagebionetworks.bridge.BridgeUtils;
+import org.sagebionetworks.bridge.Roles;
 import org.sagebionetworks.bridge.dao.SurveyDao;
 import org.sagebionetworks.bridge.exceptions.BadRequestException;
 import org.sagebionetworks.bridge.exceptions.ConstraintViolationException;
 import org.sagebionetworks.bridge.exceptions.EntityNotFoundException;
+import org.sagebionetworks.bridge.exceptions.PublishedSurveyException;
 import org.sagebionetworks.bridge.time.DateUtils;
 import org.sagebionetworks.bridge.models.ClientInfo;
 import org.sagebionetworks.bridge.models.GuidCreatedOnVersionHolder;
@@ -37,6 +38,8 @@ import org.sagebionetworks.bridge.validators.Validate;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.springframework.validation.Validator;
+
+import com.google.common.collect.ImmutableSet;
 
 @Component
 public class SurveyService {
@@ -72,26 +75,28 @@ public class SurveyService {
         this.studyService = studyService;
     }
     
+    public Survey getSurvey(Set<Roles> callerRoles, StudyIdentifier studyIdentifier, GuidCreatedOnVersionHolder keys, boolean includeElements, boolean throwException) {
+        Survey survey = surveyDao.getSurvey(keys, includeElements);
+        if (!isInStudy(callerRoles, studyIdentifier, survey)) {
+            if (throwException) {
+                throw new EntityNotFoundException(Survey.class);    
+            }
+            return null;
+        }
+        return survey;
+    }
+    
     /**
      * Get a list of all published surveys in this study, using the most recently published version of each survey.
      * These surveys will include questions (not other element types, such as info screens). Most properties beyond
      * identifiers will be removed from these surveys as they are returned in the API.
-     *
-     * @param studyIdentifier
-     * @return
      */
-    public Survey getSurvey(GuidCreatedOnVersionHolder keys, boolean includeElements) {
-        checkArgument(StringUtils.isNotBlank(keys.getGuid()), "Survey GUID cannot be null/blank");
-        checkArgument(keys.getCreatedOn() != 0L, "Survey createdOn timestamp cannot be 0");
-        
-        return surveyDao.getSurvey(keys, includeElements);
+    public Survey getSurvey(StudyIdentifier studyIdentifier, GuidCreatedOnVersionHolder keys, boolean includeElements, boolean throwException) {
+        return getSurvey(ImmutableSet.of(), studyIdentifier, keys, includeElements, throwException);
     }
 
     /**
      * Create a survey.
-     * 
-     * @param survey
-     * @return
      */
     public Survey createSurvey(Survey survey) {
         checkNotNull(survey, "Survey cannot be null");
@@ -112,12 +117,25 @@ public class SurveyService {
 
     /**
      * Update an existing survey.
-     * 
-     * @param survey
-     * @return
      */
-    public Survey updateSurvey(Survey survey) {
+    public Survey updateSurvey(StudyIdentifier studyIdentifier, Survey survey) {
         checkNotNull(survey, "Survey cannot be null");
+        
+        Survey existing = surveyDao.getSurvey(survey, false);
+        if (existing == null || (existing.isDeleted() && survey.isDeleted()) || !isInStudy(studyIdentifier, survey)) {
+            throw new EntityNotFoundException(Survey.class);
+        }
+
+        if (existing.isPublished()) {
+            // If the existing survey is published, the only thing you can do is undelete it.
+            if (existing.isDeleted() && !survey.isDeleted()) {
+                existing = surveyDao.getSurvey(survey, true); // get all the children for the update
+                existing.setDeleted(false);
+                return surveyDao.updateSurvey(existing);
+            } else {
+                throw new PublishedSurveyException(survey);
+            }
+        }
         
         Set<String> dataGroups = Collections.emptySet();
         if (survey.getStudyIdentifier() != null) {
@@ -133,35 +151,25 @@ public class SurveyService {
      * Make this version of this survey available for scheduling. One scheduled for publishing, a survey version can no
      * longer be changed (it can still be the source of a new version). There can be more than one published version of
      * a survey.
-     * 
-     * @param study
-     *            study ID of study to publish the survey to
-     * @param keys
-     *            survey keys (guid, created on timestamp)
-     * @param newSchemaRev
-     *            true if you want to cut a new survey schema, false if you should (attempt to) modify the existing one
-     * @return published survey
      */
-    public Survey publishSurvey(StudyIdentifier study, GuidCreatedOnVersionHolder keys, boolean newSchemaRev) {
-        checkArgument(StringUtils.isNotBlank(keys.getGuid()), "Survey GUID cannot be null/blank");
-        checkArgument(keys.getCreatedOn() != 0L, "Survey createdOn timestamp cannot be 0");
+    public Survey publishSurvey(StudyIdentifier studyIdentifier, GuidCreatedOnVersionHolder keys, boolean newSchemaRev) {
+        Survey existing = surveyDao.getSurvey(keys, true);
+        if (existing == null || existing.isDeleted() || !isInStudy(studyIdentifier, existing)) {
+            throw new EntityNotFoundException(Survey.class);
+        }
+        Validate.entityThrowingException(publishValidator, existing);
 
-        Survey survey = surveyDao.getSurvey(keys, true);
-        Validate.entityThrowingException(publishValidator, survey);
-
-        return surveyDao.publishSurvey(study, survey, keys, newSchemaRev);
+        return surveyDao.publishSurvey(studyIdentifier, existing, newSchemaRev);
     }
 
     /**
      * Copy the survey and return a new version of it.
-     * 
-     * @param keys
-     * @return
      */
-    public Survey versionSurvey(GuidCreatedOnVersionHolder keys) {
-        checkArgument(StringUtils.isNotBlank(keys.getGuid()), "Survey GUID cannot be null/blank");
-        checkArgument(keys.getCreatedOn() != 0L, "Survey createdOn timestamp cannot be 0");
-
+    public Survey versionSurvey(StudyIdentifier studyIdentifier, GuidCreatedOnVersionHolder keys) {
+        Survey existing = surveyDao.getSurvey(keys, false);
+        if (existing == null || existing.isDeleted() || !isInStudy(studyIdentifier, existing)) {
+            throw new EntityNotFoundException(Survey.class);
+        }
         return surveyDao.versionSurvey(keys);
     }
 
@@ -172,15 +180,11 @@ public class SurveyService {
      * survey version that could have been sent to users will remain in the API so you can look at its 
      * schema, etc. This is how study developers should delete surveys. 
      */
-    public void deleteSurvey(GuidCreatedOnVersionHolder keys) {
-        checkArgument(StringUtils.isNotBlank(keys.getGuid()), "Survey GUID cannot be null/blank");
-        checkArgument(keys.getCreatedOn() != 0L, "Survey createdOn timestamp cannot be 0");
-
+    public void deleteSurvey(StudyIdentifier studyIdentifier, GuidCreatedOnVersionHolder keys) {
         Survey existing = surveyDao.getSurvey(keys, true);
-        if (existing.isDeleted()) {
+        if (existing == null || existing.isDeleted() || !isInStudy(studyIdentifier, existing)) {
             throw new EntityNotFoundException(Survey.class);
         }
-
         // verify if a shared module refers to it
         verifySharedModuleExistence(keys);
 
@@ -199,12 +203,13 @@ public class SurveyService {
      *      right now.</li>
      * </ol>
      */
-    public void deleteSurveyPermanently(StudyIdentifier studyId, GuidCreatedOnVersionHolder keys) {
-        checkArgument(StringUtils.isNotBlank(keys.getGuid()), "Survey GUID cannot be null/blank");
-        checkArgument(keys.getCreatedOn() != 0L, "Survey createdOn timestamp cannot be 0");
-
-        checkConstraintsBeforePhysicalDelete(studyId, keys);
-
+    public void deleteSurveyPermanently(Set<Roles> callerRoles, StudyIdentifier studyIdentifier,
+            GuidCreatedOnVersionHolder keys) {
+        Survey existing = surveyDao.getSurvey(keys, false);
+        if (existing == null || !isInStudy(callerRoles, studyIdentifier, existing)) {
+            throw new EntityNotFoundException(Survey.class);
+        }
+        checkConstraintsBeforePhysicalDelete(studyIdentifier, keys);
         surveyDao.deleteSurveyPermanently(keys);
     }
 
@@ -229,11 +234,15 @@ public class SurveyService {
      * @param guid
      * @return
      */
-    public List<Survey> getSurveyAllVersions(StudyIdentifier studyIdentifier, String guid) {
+    public List<Survey> getSurveyAllVersions(StudyIdentifier studyIdentifier, String guid, boolean includeDeleted) {
         checkNotNull(studyIdentifier, Validate.CANNOT_BE_NULL, "study");
         checkArgument(isNotBlank(guid), Validate.CANNOT_BE_BLANK, "survey guid");
 
-        return surveyDao.getSurveyAllVersions(studyIdentifier, guid);
+        List<Survey> allVersions = surveyDao.getSurveyAllVersions(studyIdentifier, guid, includeDeleted);
+        if (allVersions.isEmpty()) {
+            throw new EntityNotFoundException(Survey.class);
+        }
+        return allVersions;
     }
 
     /**
@@ -247,7 +256,11 @@ public class SurveyService {
         checkNotNull(studyIdentifier, Validate.CANNOT_BE_NULL, "study");
         checkArgument(isNotBlank(guid), Validate.CANNOT_BE_BLANK, "survey guid");
 
-        return surveyDao.getSurveyMostRecentVersion(studyIdentifier, guid);
+        Survey survey = surveyDao.getSurveyMostRecentVersion(studyIdentifier, guid);
+        if (survey == null || !isInStudy(studyIdentifier, survey)) {
+            throw new EntityNotFoundException(Survey.class);
+        }
+        return survey;
     }
 
     /**
@@ -264,20 +277,21 @@ public class SurveyService {
         checkNotNull(studyIdentifier, Validate.CANNOT_BE_NULL, "study");
         checkArgument(isNotBlank(guid), Validate.CANNOT_BE_BLANK, "survey guid");
 
-        return surveyDao.getSurveyMostRecentlyPublishedVersion(studyIdentifier, guid, includeElements);
+        Survey survey = surveyDao.getSurveyMostRecentlyPublishedVersion(studyIdentifier, guid, includeElements);
+        if (survey == null || !isInStudy(studyIdentifier, survey)) {
+            throw new EntityNotFoundException(Survey.class);
+        }
+        return survey;
     }
 
     /**
      * Get the most recent version of each survey in the study that has been published. If a survey has not been
      * published, nothing is returned.
-     * 
-     * @param studyIdentifier
-     * @return
      */
-    public List<Survey> getAllSurveysMostRecentlyPublishedVersion(StudyIdentifier studyIdentifier) {
+    public List<Survey> getAllSurveysMostRecentlyPublishedVersion(StudyIdentifier studyIdentifier, boolean includeDeleted) {
         checkNotNull(studyIdentifier, Validate.CANNOT_BE_NULL, "study");
 
-        return surveyDao.getAllSurveysMostRecentlyPublishedVersion(studyIdentifier);
+        return surveyDao.getAllSurveysMostRecentlyPublishedVersion(studyIdentifier, includeDeleted);
     }
 
     /**
@@ -286,10 +300,34 @@ public class SurveyService {
      * @param studyIdentifier
      * @return
      */
-    public List<Survey> getAllSurveysMostRecentVersion(StudyIdentifier studyIdentifier) {
+    public List<Survey> getAllSurveysMostRecentVersion(StudyIdentifier studyIdentifier, boolean includeDeleted) {
         checkNotNull(studyIdentifier, Validate.CANNOT_BE_NULL, "study");
 
-        return surveyDao.getAllSurveysMostRecentVersion(studyIdentifier);
+        return surveyDao.getAllSurveysMostRecentVersion(studyIdentifier, includeDeleted);
+    }
+
+    /**
+     * Callers must operate on a survey in their own study. However our code has allowed administrators to delete 
+     * shared studies (which are not in the admin's study). For backwards compatibility, do not enforce the same 
+     * study rule for shared study surveys. Eventually we want admins to be able to switch into the shared study in 
+     * order to delete items there, then this exception to the check can be removed.
+     * 
+     * The use of an admin role to skip this check is temporary; eventually admins will be able to act on behalf 
+     * of any study, allowing them to pass all such study checks. Currently we're skipping this check for admins
+     * when they are deleting studies permanently.
+     */
+    private boolean isInStudy(Set<Roles> roles, StudyIdentifier studyId, Survey survey) {
+        if (studyId == null || roles.contains(Roles.ADMIN)) {
+            return true;
+        }
+        if (survey == null || survey.getStudyIdentifier() == null) {
+            return false;
+        }
+        return survey.getStudyIdentifier().equals(studyId.getIdentifier());
+    }
+    
+    private boolean isInStudy(StudyIdentifier studyId, Survey survey) {
+        return isInStudy(ImmutableSet.of(), studyId, survey);
     }
     
     private void checkConstraintsBeforePhysicalDelete(final StudyIdentifier studyId, final GuidCreatedOnVersionHolder keys) {
@@ -305,12 +343,14 @@ public class SurveyService {
             throwConstraintViolation(match, keys);
         }
 
-        // If there's a pointer to the published version of this study, make sure this is not the last one.
+        // If there's a pointer to the published version of this survey, make sure this is not the last published survey
         match = findFirstMatchingPlan(plans, keys, (surveyReference, theseKeys) -> {
             return surveyReference.getGuid().equals(theseKeys.getGuid());
         });
         if (match != null) {
-            long publishedSurveys = getSurveyAllVersions(studyId, keys.getGuid()).stream()
+            // A plan points to this survey's published version, so there must be at least one published version
+            // that's not logically deleted
+            long publishedSurveys = getSurveyAllVersions(studyId, keys.getGuid(), false).stream()
                     .filter(Survey::isPublished).collect(Collectors.counting());
 
             if (publishedSurveys == 1L) {
